+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-dDnWAEqa.js
@@ -1,8 +1,0 @@
-import{c as r,s as e}from"./flowDiagram-HODETNUW-CvGEVhUi.js";import{_ as a}from"./mermaid.core-XFAJFxrc.js";import"./chunk-5VM5RSS4-CH7gDIyM.js";import"./chunk-XXDRQBXY-DT8LRGEJ.js";import"./chunk-POPQ4Y6H-BQpQ1DXR.js";import"./chunk-F27PBJKO-CVRZWsl0.js";import"./channel-BVfIADSO.js";import"./index-BAYqZz8d.js";var o=a(t=>`${e(t)}
-  .swimlane.cluster rect {
-    stroke: ${t.clusterBorder} !important;
-  }
-  [data-look="neo"].cluster rect {
-    filter: none;
-  }
-`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};